Added more legal docs, Added Ares Domain, more Apollo Domain support, began naming modules in script.
Added more Legal docs to the emulator.
In doing so, learned that .dat files that have just numbers in them, serve to tell the ps3 if there is any pages that exist and how many. If the dat file doesn't exist, it defaults to one page, even so. They weren't live on the Sony domain anymore, so I added custom ones myself since I know what it wants now, and have edited the code in Legaldocs module to allow for this.
Added more Apollo pngs, added support for Ares domain.
Added solution files for visual studio.
Static-Resource Trophy-Status is broken for now!